Fix: PostService 수정 주석 추가 및 로직 수정, 예외 처리 Related to: #S12P11A605-99

+++ collabit-server/src/main/java/com/collabit/community/service/PostService.java
@@ -1,5 +1,6 @@
 package com.collabit.community.service;
 
+import com.collabit.community.exception.PostNotFoundException;
 import com.collabit.global.service.S3Service;
 import com.collabit.community.domain.dto.CreatePostRequestDTO;
 import com.collabit.community.domain.dto.CreatePostResponseDTO;
@@ -8,12 +9,9 @@ import com.collabit.community.domain.dto.UpdatePostRequestDTO;
 import com.collabit.community.domain.entity.Image;
 import com.collabit.community.domain.entity.Post;
 import com.collabit.community.repository.ImageRepository;
-import com.collabit.community.repository.PostLikeRepository;
 import com.collabit.community.repository.PostRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.cache.annotation.Cacheable;
-import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -27,59 +25,79 @@ public class PostService {
     private final S3Service s3Service;
     private final PostRepository postRepository;
     private final ImageRepository imageRepository;
-    private final PostCacheService postCacheService;
+    private final LikeCacheService likeCacheService;
 
     private static final String DIR_NAME = "posts";
 
     @Transactional
     public CreatePostResponseDTO createPost(String userCode, CreatePostRequestDTO request) {
+        // RequestDTO로 post 생성
         Post post = Post.builder()
             .userCode(userCode)
             .content(request.getContent())
             .build();
         Post savedPost = postRepository.save(post);
+        
+        // Image가 있으면 image 업로드 및 DB저장
+        if(request.getImages() != null) {
+            Arrays.stream(request.getImages())
+                .map(file -> {
+                    String url = s3Service.upload(file, DIR_NAME); // S3에 업로드 후 URL 반환
+                    return Image.builder()
+                        .url(url)
+                        .post(savedPost)
+                        .build(); // 반환된 URL을 이용해 Image 객체 생성
+                })
+                .forEach(imageRepository::save);
+        }
 
-        Arrays.stream(request.getImages())
-            .map(file -> {
-                String url = s3Service.upload(file, DIR_NAME); // S3에 업로드 후 URL 반환
-                return Image.builder()
-                    .url(url)
-                    .post(savedPost)
-                    .build(); // 반환된 URL을 이용해 Image 객체 생성
-            })
-            .forEach(imageRepository::save);
-
+        // savedPost로 responseDTO 생성 후 반환
         CreatePostResponseDTO responseDTO = CreatePostResponseDTO.builder()
             .code(savedPost.getCode()).build();
         return responseDTO;
     }
 
     public List<GetPostResponseDTO> getPostList(String userCode) {
+        // 게시글 목록을 담을 List
         List<GetPostResponseDTO> list = new ArrayList<>();
-
+        
+        // 전체 Post 조회
         List<Post> posts = postRepository.findAll();
         for (Post post : posts) {
+            // post들로 responseDTO 생성
             GetPostResponseDTO responseDTO = buildDTO(post,userCode);
+            // list에 추가
             list.add(responseDTO);
         }
+        // list 반환
         return list;
     }
 
     public GetPostResponseDTO getPost(String userCode, int postCode) {
+        // postCode에 해당하는 post 조회
         Post post = postRepository.findByCode(postCode);
+        // 해당 post가 없으면 예외 처리
+        if(post==null) throw new PostNotFoundException();
+        // responseDTO 생성
         GetPostResponseDTO responseDTO = buildDTO(post,userCode);
-
+        // 반환
         return responseDTO;
     }
 
     @Transactional
     public GetPostResponseDTO updatePost(String userCode, int postCode, UpdatePostRequestDTO requestDTO) {
-        Post post = postRepository.findByCode(postCode); // Post를 가져옴
-        post.setContent(requestDTO.getContent()); // Post 내용 업데이트
-
-        String[] imageUrls = requestDTO.getImages(); // 삭제할 이미지 URL 배열
-        List<Image> postImages = post.getImages(); // Post에 연결된 이미지 리스트 가져오기
-
+        // postCode에 해당하는 post 조회
+        Post post = postRepository.findByCode(postCode);
+        // 해당 post가 없으면 예외 처리
+        if(post==null) throw new PostNotFoundException();
+        // 가지고 온 post의 userCode와 요청 userCode가 일치하지 않으면 예외 처리
+        if(!post.getUserCode().equals(userCode)) throw new RuntimeException();
+        // Post 내용 업데이트
+        post.setContent(requestDTO.getContent());
+        // 삭제할 이미지 URL 배열
+        String[] imageUrls = requestDTO.getImages();
+        // Post에 연결된 이미지 리스트 가져오기
+        List<Image> postImages = post.getImages();
         // 삭제할 URL에 해당하는 이미지 삭제
         postImages.removeIf(image -> {
             boolean shouldRemove = Arrays.asList(imageUrls).contains(image.getUrl());
@@ -87,37 +105,53 @@ public class PostService {
                 // S3에서 이미지 삭제
                 s3Service.delete(image.getUrl());
             }
-            return shouldRemove; // 삭제 조건에 해당하면 리스트에서 제거
+            // 삭제 조건에 해당하면 리스트에서 제거
+            return shouldRemove;
         });
-
-        Post savedPost = postRepository.save(post);
-
-        GetPostResponseDTO responseDTO = buildDTO(savedPost,userCode);
+        
+        // 바뀐 Post update
+        Post updatedPost = postRepository.save(post);
+        
+        // updatedPost와 userCode로 responseDTO 생성
+        GetPostResponseDTO responseDTO = buildDTO(updatedPost,userCode);
 
         return responseDTO;
     }
 
     @Transactional
-    public void deletePost(int postCode) {
+    public void deletePost(String userCode, int postCode) {
+        // postCode에 해당하는 post 조회
         Post post = postRepository.findByCode(postCode);
-        // List<Image> images = imageRepository.findByPost(post);
+        // 해당 post가 없으면 예외 처리
+        if(post==null) throw new PostNotFoundException();
+        // 가지고 온 post의 userCode와 요청 userCode가 일치하지 않으면 예외 처리
+        if(!post.getUserCode().equals(userCode)) throw new RuntimeException();
+        // Post에 연결된 이미지 리스트 가져오기
         List<Image> images = post.getImages();
+        // 삭제할 URL에 해당하는 이미지 삭제
         String[] imageUrls = images.stream()
             .map(Image::getUrl)
             .toArray(String[]::new);
         for (String url : imageUrls) {
             s3Service.delete(url);
         }
+        // 해당 post 삭제
         postRepository.deleteByCode(postCode);
     }
-
+    
+    // Post와 userCode로 responseDTO 만드는 메서드
     public GetPostResponseDTO buildDTO (Post post, String userCode) {
+        // Post에 연결된 이미지 리스트 가져오기
         List<Image> images = post.getImages();
+        // 이미지에서 url 가져오기
         String[] imageUrls = images.stream()
             .map(Image::getUrl)
             .toArray(String[]::new);
-        int likeCount = postCacheService.getLikeCount((post.getCode()));
-        boolean isLiked = postCacheService.getIsLiked(userCode, post.getCode());
+        // likeCount를 Look Aside 전략으로 조회
+        int likeCount = likeCacheService.getLikeCount((post.getCode()));
+        // isLiked를 Look Aside 전략으로 조회
+        boolean isLiked = likeCacheService.getIsLiked(userCode, post.getCode());
+        // responseDTO를 생성 후 반환
         return GetPostResponseDTO.builder()
             .code(post.getCode())
             .userCode(post.getUserCode())
